fix: close the remaining code-review findings and keep the stack warning-free
Authorize clan and room mutations, make the shared session state thread-safe,
and close the remaining data-integrity gaps flagged by the review:

- Clan write handlers now resolve the caller's membership row, require the
  leader for leadership and emblem-editor changes, and record the caller's
  membership row as the notice author instead of its character id.
- SetGame/UpdatePings/StartRound require the caller to host the room.
- Remove the dead, unsynchronized TcpServerBase session dictionary; guard
  LobbyTrackerService with a lock and debounce the per-disconnect republish.
- Reserve outbound sequence numbers with Interlocked.Increment before encoding.
- Add the sessions.user_id unique index; serialize registration, session and
  mail writes so the check-then-insert races cannot duplicate or overshoot.
- Throw from EncodePacket when a payload exceeds the protocol limit, keep a
  matched automatch searcher so a late cancel reports TooLate, and take the
  gameplay server account credentials from configuration (random password when
  unset, never logged).
- Fix the nits: coalesce(MAX(id),1) for setval, parameterized round-roster
  insert, single-snapshot friends list, MemoryStream accumulation buffer with a
  separate read cursor, fixed-size asserts, long notice time, and a correct
  inbound-log command label.
- Add frame-cipher, handshake, packet-codec and malformed-DNS tests.

Verified with a clean Release build (0 warnings), 19 passing tests, and a local
docker compose run: TCP traffic logs at Debug with no no-handler warning and no
runtime warnings or errors.

+++ src/Shared/Domain/Clans/ClanService.Membership.cs
@@ -159,7 +159,7 @@ public sealed partial class ClanService
         CancellationToken cancellationToken = default)
     {
         await using var context = await CreateContextAsync(cancellationToken);
-        var secondsSinceEpoch = (int)DateTimeOffset.UtcNow.ToUnixTimeSeconds();
+        var secondsSinceEpoch = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
 
         await context.Clans
             .Where(clan => clan.Identifier == clanIdentifier)
